Step4 로또(수동) 피드백 (#3219)
* Feat(Lotto) : Step3 로또(2등) 피드백

- 명확한 메소드, 상수 네이밍 고려

* Feat(Lotto) : Step3 로또(2등) 피드백

- 불필요한 메소드 분리를 통한 혼동 제거

* Feat(Lotto) : Step3 로또(2등) 피드백

- 주석 처리된 테스트 코드 원복

* Feat(Lotto) : Step3 로또(2등) 피드백

- 클래스명 변경 -> 명확한 역할 구분

* Feat(Lotto) : Step3 로또(2등) 피드백

- 클래스명 변경 -> 명확한 역할 구분
- 로또 번호 생성 클래스 분리

* Feat(Lotto) : Step3 로또(2등) 피드백

- 클래스명 변경 -> 명확한 역할 구분
- 로또 번호 생성 클래스 분리

* Feat(Lotto) : Step3 로또(2등) 피드백

- 의도에 맞는 메소드 네이밍 작성
- BonusBall 에대한 간접 참조 변경
- 테스트 코드 변경

* Feat(Lotto) : Step3 로또(2등) 피드백

- 로또 1등 당첨이 1명 이상인 경우에 대한 Integer 에서 Long 타입으로 변경

* Feat(Lotto) : Step3 로또(2등) 피드백

- 뷰에 의존하는 로직에 대한 고려 및 수정
- RewardTable 에 예외 등수 MISS 추가

* Feat(Lotto) : Step3 로또(2등) 피드백

- 뷰에 의존하는 로직에 대한 고려 및 수정
- RewardTable 에 예외 등수 MISS 추가

* Feat(Lotto) : Step4 로또(수동) 개발

- 기능 요구 사항 작성

* Feat(Lotto) : Step4 로또(수동) 개발

- 수동 구입 메시지 개발

* Feat(Lotto) : Step4 로또(수동) 개발

- 로또 수동 구입 관련 로직 개발

* Feat(Lotto) : Step4 로또(수동) 개발

- LottoNumber 객체 적용
- 테스트코드 수정
- 중복 상수 선언

* Feat(Lotto) : Step4 로또(수동) 피드백

- 불필요한 매직넘버 사용 자제하기

* Feat(Lotto) : Step4 로또(수동) 피드백

- 메소드에서만 사용하는 불필요한 필드(파라미터)제거

* Feat(Lotto) : Step4 로또(수동) 피드백

- 동시성에 대한 고려

* Feat(Lotto) : Step4 로또(수동) 피드백

- 박싱된 기본 타입보다는 기본 타입을 사용

* Feat(Lotto) : Step4 로또(수동) 피드백

- 메소드 네이밍 신중하게 고려하기

* Feat(Lotto) : Step4 로또(수동) 피드백

- 메소드 네이밍 신중하게 고려하기

* Feat(Lotto) : Step4 로또(수동) 피드백

- 오타 수정

* Feat(Lotto) : Step4 로또(수동) 피드백

- RewardTable bonus 필드 제거

* Feat(Lotto) : Step4 로또(수동) 피드백

- 같은 값 재활용에 대해서도 같은 역할인지 고려하기

* Feat(Lotto) : Step4 로또(수동) 피드백

- 상수를 관리하는 클래스라면 private 생성자를 통해 방어하기
- 상수의 역할에 맞게 공통 분리, 다르면 세분화하여 관리

+++ src/test/java/lotto/domian/RewardTest.java
@@ -15,11 +15,11 @@ public class RewardTest {
     @DisplayName("로또 게임 맞춘 전체 금액을 응답한다.")
     @Test
     public void rewardTest_게임_정답_전체금액_테스트() {
-        Map<RewardTable, Integer> resultGameMap = new HashMap<>();
-        resultGameMap.put(RewardTable.of(3, null), 2);    //  10_000
-        resultGameMap.put(RewardTable.of(4, null), 2);    //  100_000
-        resultGameMap.put(RewardTable.of(5, false), 1);   //  1_500_000
-        resultGameMap.put(RewardTable.of(6, null), 1);    //  2_000_000_000
+        Map<RewardTable, Long> resultGameMap = new HashMap<>();
+        resultGameMap.put(RewardTable.of(3, null), 2L);    //  10_000
+        resultGameMap.put(RewardTable.of(4, null), 2L);    //  100_000
+        resultGameMap.put(RewardTable.of(5, false), 1L);   //  1_500_000
+        resultGameMap.put(RewardTable.of(6, null), 1L);    //  2_000_000_000
 
         assertThat(new Reward().sumTotalMatchPrice(resultGameMap)).isEqualTo(2_001_610_000);
     }
